requirements.txt
discord.py==1.7.3
asyncpraw==7.5.0
python-dotenv==0.20.0